# docs/swagger.yaml
basePath: /api
definitions:
  github_com_torvictorvic_seek_internal_domain.Candidate:
    properties:
      created_at:
        type: string
      email:
        type: string
      gender:
        type: string
      id:
        type: integer
      name:
        type: string
      salary_expected:
        type: number
      updated_at:
        type: string
    type: object
host: localhost:8080
info:
  contact:
    email: soporte@ccccc.com
    name: Soporte
    url: http://www.ccccc.com/support
  description: Esta API gestiona candidatos en un proceso de reclutamiento
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://swagger.io/terms/
  title: Sistema de Gestión de Candidatos
  version: "1.0"
paths:
  /candidates:
    get:
      consumes:
      - application/json
      description: Retorna la lista de todos los candidatos
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/github_com_torvictorvic_seek_internal_domain.Candidate'
            type: array
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
      security:
      - Bearer: []
      summary: Listar todos los candidatos
      tags:
      - Candidates
    post:
      consumes:
      - application/json
      description: Crea un candidato con los datos enviados en el body
      parameters:
      - description: Datos del candidato
        in: body
        name: candidate
        required: true
        schema:
          $ref: '#/definitions/github_com_torvictorvic_seek_internal_domain.Candidate'
      produces:
      - application/json
      responses:
        "200":
          description: ok
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      security:
      - Bearer: []
      summary: Crear un nuevo candidato
      tags:
      - Candidates
  /candidates/{id}:
    delete:
      consumes:
      - application/json
      description: Borra un candidato cuyo ID se pasa como parámetro
      parameters:
      - description: ID del Candidato
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_torvictorvic_seek_internal_domain.Candidate'
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Candidato no encontrado
          schema:
            additionalProperties: true
            type: object
      security:
      - Bearer: []
      summary: Borra un candidato por ID
      tags:
      - Candidates
    get:
      consumes:
      - application/json
      description: Retorna el candidato cuyo ID se pasa como parámetro
      parameters:
      - description: ID del Candidato
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_torvictorvic_seek_internal_domain.Candidate'
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Candidato no encontrado
          schema:
            additionalProperties: true
            type: object
      security:
      - Bearer: []
      summary: Obtener candidato por ID
      tags:
      - Candidates
    put:
      consumes:
      - application/json
      description: Actualiza un candidato con los datos enviados en el body
      parameters:
      - description: ID del Candidato
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: ok
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      security:
      - Bearer: []
      summary: Actualiza un candidato
      tags:
      - Candidates
swagger: "2.0"